streamlit/requirements.txt
Pillow==9.2.0
numpy==1.22.4
requests==2.28.1
streamlit==1.22.0
tensorflow==2.9.1
